refactor(theme-configurator): enhance theme configurator with preview manager integration
- Removed legacy functions and constants related to preview management, streamlining the codebase.
- Introduced a new `preview-manager.ts` to handle preview rendering and controller management, improving modularity.
- Updated theme configurator to utilize the new preview manager, enhancing theme preview functionality and reducing code duplication.
- Re-exported test helpers from the preview manager for better accessibility in the theme configurator.
- Improved overall structure and readability of the theme configurator code.
